add paged list of digital assets

+++ src/app/schemas/digital_asset/digital_asset_schemas.ts
@@ -103,7 +103,7 @@ export const digitalAssetCounterSchema = {
 	required: ["counter"],
 	properties: {
 	  counter: {
-		dataType: "uint64",
+		dataType: "uint32",
 		fieldNumber: 1,
 	  },
 	},
